move components to UI
